Server1 (#8)
* Added schema and user model

* Added schema to create user

* Login and signup connected to mongoose

* Added tokens

* Added city schema

* Added helmet

* Fixed name

* Connected CRUD operations with backend

* Added patch and delete operations

---------

+++ src/contexts/CitiesContext.jsx
@@ -28,6 +28,12 @@ function reducer(state, action) {
         isLoading: false,
         cities: action.payload,
       };
+    case "cities/update":
+      return {
+        ...state,
+        isLoading: false,
+        cities: action.payload,
+      };
 
     case "city/loaded":
       return { ...state, isLoading: false, currentCity: action.payload };
@@ -44,7 +50,7 @@ function reducer(state, action) {
       return {
         ...state,
         isLoading: false,
-        cities: state.cities.filter((city) => city.id !== action.payload),
+        cities: state.cities.filter((city) => city._id !== action.payload),
         currentCity: {},
       };
 
@@ -60,6 +66,15 @@ function reducer(state, action) {
   }
 }
 
+function verifyToken() {
+  const token = localStorage.getItem("authToken");
+
+  return {
+    "Content-Type": "application/json",
+    Authorization: token ? `Bearer ${token}` : "",
+  };
+}
+
 function CitiesProvider({ children }) {
   const [{ cities, isLoading, currentCity, error }, dispatch] = useReducer(
     reducer,
@@ -70,10 +85,13 @@ function CitiesProvider({ children }) {
     async function fetchCities() {
       dispatch({ type: "loading" });
 
+      const headers = verifyToken();
+
       try {
-        const res = await fetch(`${BASE_URL}/cities`);
+        const res = await fetch(`${BASE_URL}/app/cities`, { headers });
         const data = await res.json();
-        dispatch({ type: "cities/loaded", payload: data });
+        console.log(data);
+        dispatch({ type: "cities/loaded", payload: data.cities });
       } catch {
         dispatch({
           type: "rejected",
@@ -90,14 +108,20 @@ function CitiesProvider({ children }) {
 
       dispatch({ type: "loading" });
 
+      const headers = verifyToken();
+
       try {
-        const res = await fetch(`${BASE_URL}/cities/${id}`);
+        const res = await fetch(`${BASE_URL}/app/cities/${id}`, { headers });
+        if (!res.ok) {
+          throw new Error("Failed to fetch city"); // Custom error message
+        }
+
         const data = await res.json();
         dispatch({ type: "city/loaded", payload: data });
-      } catch {
+      } catch (error) {
         dispatch({
           type: "rejected",
-          payload: "There was an error loading the city...",
+          payload: error.message || "There was an error loading the city...",
         });
       }
     },
@@ -107,17 +131,45 @@ function CitiesProvider({ children }) {
   async function createCity(newCity) {
     dispatch({ type: "loading" });
 
+    const headers = verifyToken();
+
     try {
-      const res = await fetch(`${BASE_URL}/cities`, {
+      const res = await fetch(`${BASE_URL}/app/cities`, {
         method: "POST",
-        body: JSON.stringify(newCity),
         headers: {
-          "Content-Type": "application/json",
+          ...headers,
         },
+        body: JSON.stringify(newCity),
       });
+
       const data = await res.json();
 
-      dispatch({ type: "city/created", payload: data });
+      dispatch({ type: "city/created", payload: data.city });
+    } catch {
+      dispatch({
+        type: "rejected",
+        payload: "There was an error creating the city...",
+      });
+    }
+  }
+
+  async function updateCity(id, newData) {
+    dispatch({ type: "loading" });
+
+    const headers = verifyToken();
+
+    try {
+      const res = await fetch(`${BASE_URL}/app/cities/${id}`, {
+        method: "PATCH",
+        headers: {
+          ...headers,
+        },
+        body: JSON.stringify(newData),
+      });
+
+      const data = await res.json();
+
+      dispatch({ type: "city/update", payload: data.city });
     } catch {
       dispatch({
         type: "rejected",
@@ -127,18 +179,28 @@ function CitiesProvider({ children }) {
   }
 
   async function deleteCity(id) {
+    if (Number(id) === currentCity.id) return;
+
     dispatch({ type: "loading" });
 
+    const headers = verifyToken();
+
     try {
-      await fetch(`${BASE_URL}/cities/${id}`, {
+      const deleteResponse = await fetch(`${BASE_URL}/app/cities/${id}`, {
         method: "DELETE",
+        headers,
       });
 
+      if (!deleteResponse.ok) {
+        throw new Error("Failed to delete city");
+      }
+
+      // Directly update the local state
       dispatch({ type: "city/deleted", payload: id });
-    } catch {
+    } catch (error) {
       dispatch({
         type: "rejected",
-        payload: "There was an error deleting the city...",
+        payload: error.message || "There was an error deleting the city...",
       });
     }
   }
@@ -153,6 +215,7 @@ function CitiesProvider({ children }) {
         getCity,
         createCity,
         deleteCity,
+        updateCity,
       }}
     >
       {children}
